refactor: Move helpers.js in separate files
Different helpers were in root cozy-client folder.
I separated them in different helper files in an helpers folder.

No API changes.
