+++ packages/cozy-client/src/helpers/index.js
@@ -1,0 +1,10 @@
+export {
+  deconstructCozyWebLinkWithSlug,
+  generateWebLink,
+  ensureFirstSlash,
+  rootCozyUrl,
+  InvalidCozyUrlError,
+  InvalidProtocolError
+} from './urlHelper'
+
+export { dehydrate } from './dehydrateHelper'
